actor,category,city,film,country,address CRUR operations

+++ src/main/java/gov/iti/jets/model/dtos/CategoryDto.java
@@ -17,7 +17,7 @@ import java.util.*;
 @ToString
 public class CategoryDto {
 
-  private Byte categoryId;
+  private Short categoryId;
   private String name;
   private Timestamp lastUpdate;
 }
